Changed tag to UInt64 to better align with TKBERTLVRecord. BytesRead is now always set to 0 if we fail parsing a TLV.

+++ YubiKit/YubiKit/Helpers/YKFTLVRecord.h
@@ -16,7 +16,7 @@
 #define TLVRecord_h
 
 
-typedef UInt32 YKFTLVTag;
+typedef UInt64 YKFTLVTag;
 
 @interface YKFTLVRecord : NSObject
